Refactor [#9] 홈 메인 뷰 분리
